backend/requirements.txt
fastapi
uvicorn
httpx
python-multipart
python-dotenv
pydantic>=2.7
pydantic-settings
beautifulsoup4
sentence-transformers
langgraph
qdrant-client
fastembed
structlog
opentelemetry-api
